refactor: Updated folders structure
