+++ src/constants/currencies.ts
