+++ src/entities/index.ts
@@ -1,10 +1,8 @@
-export * from './baseCurrency'
-export * from './currency'
-export * from './ether'
-export * from './nativeCurrency'
-export * from './pair'
-export * from './route'
-export * from './token'
-export * from './trade'
-export * from './currencies'
-export * from './fractions'
+export * from 'entities/baseCurrency'
+export * from 'entities/currency'
+export * from 'entities/nativeCurrency'
+export * from 'entities/pair'
+export * from 'entities/route'
+export * from 'entities/token'
+export * from 'entities/trade'
+export * from 'entities/fractions'
