+++ src/index.ts
@@ -1,8 +1,18 @@
 import JSBI from 'jsbi'
 export { JSBI }
 
-export { BigintIsh, ChainId, TradeType, Rounding, FACTORY_ADDRESS, INIT_CODE_HASH, MINIMUM_LIQUIDITY } from 'constants'
-
+export {
+  BigintIsh,
+  ChainId,
+  TradeType,
+  Rounding,
+  FACTORY_ADDRESS,
+  INIT_CODE_HASH,
+  MINIMUM_LIQUIDITY,
+  NATIVE_CURRENCY,
+  WRAPPED_NATIVE_CURRENCY,
+  WETH,
+} from 'constants'
 export * from 'errors'
 export * from 'entities'
 export * from 'router'
