+++ test/router.test.ts
@@ -1,6 +1,6 @@
 import invariant from 'tiny-invariant'
 import { ChainId, CurrencyAmount, Pair, Percent, Route, Router, Token, TokenAmount, Trade } from '../src'
-import { NATIVE_CURRENCY, WRAPPED_NATIVE_CURRENCY } from '../src/entities/currencies'
+import { NATIVE_CURRENCY, WRAPPED_NATIVE_CURRENCY } from '../src/constants'
 import JSBI from 'jsbi'
 
 function checkDeadline(deadline: string[] | string): void {
